Change lib files. Added files for binnary graph. Removed files for list nodes.

+++ AdditionalFiles/GraphLib.c
@@ -6,10 +6,10 @@
 
 GRAPH initEmptyGraph(int nodeCount) {
 	GRAPH graph;
-	graph = calloc(nodeCount, sizeof(short int*));
+	graph = calloc(nodeCount, sizeof(int*));
 	for (size_t i = 0; i < nodeCount; i++)
 	{
-		graph[i] = calloc(nodeCount, sizeof(short int));
+		graph[i] = calloc(nodeCount, sizeof(int));
 		for (size_t j = 0; j < nodeCount; j++)
 		{
 			graph[i][j] = 0;
@@ -20,7 +20,7 @@ GRAPH initEmptyGraph(int nodeCount) {
 
 int getGraphLength(GRAPH graph) {
 	int sqrLength;
-	sqrLength = sizeof(graph[0]) / sizeof(short int);
+	sqrLength = sizeof(graph[0]) / sizeof(int);
 	return sqrt(sqrLength);
 }
 
@@ -39,10 +39,10 @@ GRAPH addNodes(GRAPH graph, int nodeCount) {
 	if (resLength < graphLength) {
 		resLength = INT_MAX;
 	}
-	graph = realloc(graph, resLength * sizeof(short int*));
+	graph = realloc(graph, resLength * sizeof(int*));
 	for (size_t i = 0; i < graphLength; i++)
 	{
-		graph[i] = realloc(graph[i], resLength * sizeof(short int));
+		graph[i] = realloc(graph[i], resLength * sizeof(int));
 		for (size_t j = graphLength; j < resLength; j++)
 		{
 			graph[i][j] = 0;
@@ -50,7 +50,7 @@ GRAPH addNodes(GRAPH graph, int nodeCount) {
 	}
 	for (size_t i = graphLength; i < resLength; i++)
 	{
-		graph[i] = realloc(graph[i], resLength * sizeof(short int));
+		graph[i] = realloc(graph[i], resLength * sizeof(int));
 		for (size_t j = 0; j < resLength; j++)
 		{
 			graph[i][j] = 0;
@@ -65,10 +65,10 @@ GRAPH removeNodes(GRAPH graph, int nodeCount) {
 	GRAPH newGraph;
 	newGraph = NULL;
 	if (resLength > 0) {
-		newGraph = calloc(resLength, sizeof(short int*));
+		newGraph = calloc(resLength, sizeof(int*));
 		for (size_t i = 0; i < resLength; i++)
 		{
-			newGraph[i] = calloc(resLength, sizeof(short int));
+			newGraph[i] = calloc(resLength, sizeof(int));
 			for (size_t j = 0; j < resLength; j++)
 			{
 				newGraph[i][j] = graph[i][j];
